misc update

+++ v2/pkg/subscraping/agent.go
@@ -46,7 +46,7 @@ func NewSession(domain string, proxy string, rateLimit, timeout int) (*Session, 
 
 	// Initiate rate limit instance
 	if rateLimit > 0 {
-		session.RateLimiter = ratelimit.New(context.Background(), int64(rateLimit), time.Second)
+		session.RateLimiter = ratelimit.New(context.Background(), uint(rateLimit), time.Second)
 	} else {
 		session.RateLimiter = ratelimit.NewUnlimited(context.Background())
 	}
